类型

+++ ucs-service/ucs-enterprise/src/main/java/com/sailmi/enterprise/mapper/EnterpriseMapper.java
@@ -81,5 +81,5 @@ public interface EnterpriseMapper extends BaseMapper<Enterprise> {
 	 * @Author: syt
 	 * @Date: 2020/10/30/0030 17:41
 	 */
-	int joinEnterprise(@Param("id") BigInteger id, @Param("userId") BigInteger userId);
+	int joinEnterprise(@Param("id") BigInteger id, @Param("userId") Long userId);
 }
